fix: resolve Week 1 review blockers

+++ src/main/java/com/payflow/api/dto/response/WalletResponse.java
@@ -3,7 +3,6 @@ package com.payflow.api.dto.response;
 import com.payflow.domain.model.wallet.Wallet;
 import com.payflow.domain.model.wallet.WalletStatus;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.Currency;
 import java.util.UUID;
@@ -11,7 +10,7 @@ import java.util.UUID;
 public record WalletResponse(
         UUID id,
         Currency currency,
-        BigDecimal balance,
+        Long balance,
         WalletStatus status,
         Instant createdAt
 ) {
